Move all triton related testing utils into shared file
[ghstack-poisoned]
